# project/backend/billing/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"eats/backend/billing/domain"
	"eats/backend/common/shared"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	"github.com/shopspring/decimal"
)

// Defines values for DocumentType.
const (
	Receipt DocumentType = "receipt"
)

// Defines values for TaxType.
const (
	Gst      TaxType = "gst"
	SalesTax TaxType = "sales-tax"
	Vat      TaxType = "vat"
)

// Address defines model for Address.
type Address struct {
	// City City name
	City string `json:"city"`

	// CountryCode ISO 3166-1 alpha-2 country code
	CountryCode shared.CountryCode `json:"country_code"`

	// Line1 First address line
	Line1 string `json:"line1"`

	// Line2 Second address line
	Line2 string `json:"line2"`

	// PostalCode Postal code
	PostalCode string `json:"postal_code"`
}

// CreateDocument defines model for CreateDocument.
type CreateDocument struct {
	Buyer LegalEntity `json:"buyer"`

	// Currency Currency code in ISO 4217 format
	Currency Currency `json:"currency"`

	// IssueDate Document issue date and time
	IssueDate time.Time `json:"issue_date"`

	// LineItems Document line items
	LineItems []LineItem  `json:"line_items"`
	Seller    LegalEntity `json:"seller"`
}

// Currency Currency code in ISO 4217 format
type Currency = shared.Currency

// Decimal defines model for Decimal.
type Decimal = decimal.Decimal

// DocumentResponse defines model for DocumentResponse.
type DocumentResponse struct {
	Buyer LegalEntity `json:"buyer"`

	// Currency Currency code in ISO 4217 format
	Currency Currency `json:"currency"`

	// DocumentNumber Human-readable document number
	DocumentNumber string `json:"document_number"`

	// DocumentType Type of the billing document
	DocumentType DocumentType          `json:"document_type"`
	IssueDate    time.Time             `json:"issue_date"`
	LineItems    []ResponseLineItem    `json:"line_items"`
	Seller       LegalEntity           `json:"seller"`
	Summary      PriceBreakdownSummary `json:"summary"`

	// Uuid UUID of a document
	Uuid DocumentUUID `json:"uuid"`
}

// DocumentType Type of the billing document
type DocumentType string

// DocumentUUID UUID of a document
type DocumentUUID = domain.DocumentUUID

// ErrorDetails defines model for ErrorDetails.
type ErrorDetails struct {
	EntityId   *string `json:"entity_id,omitempty"`
	EntityType *string `json:"entity_type,omitempty"`
	ErrorSlug  string  `json:"error_slug"`
	Message    string  `json:"message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Details Detailed error information
	Details []ErrorDetails `json:"details"`

	// Message Human-readable error message
	Message string `json:"message"`

	// Slug Error slug
	Slug string `json:"slug"`
}

// LegalEntity defines model for LegalEntity.
type LegalEntity struct {
	Address Address `json:"address"`

	// Name Legal entity name
	Name string `json:"name"`

	// TaxId Tax identification number
	TaxId *TaxID `json:"tax_id,omitempty"`
}

// LineItem defines model for LineItem.
type LineItem struct {
	IsGross bool `json:"is_gross"`

	// Name Item description/name
	Name string `json:"name"`

	// Quantity Quantity of items
	Quantity   int     `json:"quantity"`
	UnitAmount Decimal `json:"unit_amount"`
}

// PriceBreakdown defines model for PriceBreakdown.
type PriceBreakdown struct {
	GrossAmount Decimal `json:"gross_amount"`
	NetAmount   Decimal `json:"net_amount"`
	TaxAmount   Decimal `json:"tax_amount"`
	TaxRate     Decimal `json:"tax_rate"`

	// TaxType Type of tax applied
	TaxType         TaxType `json:"tax_type"`
	UnitGrossAmount Decimal `json:"unit_gross_amount"`
	UnitNetAmount   Decimal `json:"unit_net_amount"`
	UnitTaxAmount   Decimal `json:"unit_tax_amount"`
}

// PriceBreakdownSummary defines model for PriceBreakdownSummary.
type PriceBreakdownSummary struct {
	GrossAmount Decimal      `json:"gross_amount"`
	NetAmount   Decimal      `json:"net_amount"`
	TaxAmount   Decimal      `json:"tax_amount"`
	Taxes       []TaxSummary `json:"taxes"`
}

// ResponseLineItem defines model for ResponseLineItem.
type ResponseLineItem struct {
	Breakdown PriceBreakdown `json:"breakdown"`
	Name      string         `json:"name"`
	Quantity  int            `json:"quantity"`
}

// TaxID Tax identification number
type TaxID = shared.TaxID

// TaxSummary defines model for TaxSummary.
type TaxSummary struct {
	NetAmount Decimal `json:"net_amount"`
	TaxAmount Decimal `json:"tax_amount"`
	TaxRate   Decimal `json:"tax_rate"`

	// TaxType Type of tax applied
	TaxType TaxType `json:"tax_type"`
}

// TaxType Type of tax applied
type TaxType string

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorResponse

// CreateReceiptJSONRequestBody defines body for CreateReceipt for application/json ContentType.
type CreateReceiptJSONRequestBody = CreateDocument

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create a new receipt
	// (POST /billing/documents/receipt)
	CreateReceipt(ctx echo.Context) error
	// Get a document by UUID
	// (GET /billing/documents/{document_uuid})
	GetDocument(ctx echo.Context, documentUuid DocumentUUID) error
	// Print a document as HTML
	// (POST /billing/documents/{document_uuid}/print)
	PrintDocument(ctx echo.Context, documentUuid DocumentUUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CreateReceipt converts echo context to params.
func (w *ServerInterfaceWrapper) CreateReceipt(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateReceipt(ctx)
	return err
}

// GetDocument converts echo context to params.
func (w *ServerInterfaceWrapper) GetDocument(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "document_uuid" -------------
	var documentUuid DocumentUUID

	err = runtime.BindStyledParameterWithOptions("simple", "document_uuid", ctx.Param("document_uuid"), &documentUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter document_uuid: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetDocument(ctx, documentUuid)
	return err
}

// PrintDocument converts echo context to params.
func (w *ServerInterfaceWrapper) PrintDocument(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "document_uuid" -------------
	var documentUuid DocumentUUID

	err = runtime.BindStyledParameterWithOptions("simple", "document_uuid", ctx.Param("document_uuid"), &documentUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter document_uuid: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PrintDocument(ctx, documentUuid)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/billing/documents/receipt", wrapper.CreateReceipt)
	router.GET(baseURL+"/billing/documents/:document_uuid", wrapper.GetDocument)
	router.POST(baseURL+"/billing/documents/:document_uuid/print", wrapper.PrintDocument)

}

type BadRequestJSONResponse ErrorResponse

type NotFoundJSONResponse ErrorResponse

type UnauthorizedJSONResponse ErrorResponse

type CreateReceiptRequestObject struct {
	Body *CreateReceiptJSONRequestBody
}

type CreateReceiptResponseObject interface {
	VisitCreateReceiptResponse(w http.ResponseWriter) error
}

type CreateReceipt201JSONResponse struct {
	// DocumentUuid UUID of a document
	DocumentUuid DocumentUUID `json:"document_uuid"`
}

func (response CreateReceipt201JSONResponse) VisitCreateReceiptResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateReceipt400JSONResponse struct{ BadRequestJSONResponse }

func (response CreateReceipt400JSONResponse) VisitCreateReceiptResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateReceipt401JSONResponse struct{ UnauthorizedJSONResponse }

func (response CreateReceipt401JSONResponse) VisitCreateReceiptResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type GetDocumentRequestObject struct {
	DocumentUuid DocumentUUID `json:"document_uuid"`
}

type GetDocumentResponseObject interface {
	VisitGetDocumentResponse(w http.ResponseWriter) error
}

type GetDocument200JSONResponse DocumentResponse

func (response GetDocument200JSONResponse) VisitGetDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetDocument400JSONResponse struct{ BadRequestJSONResponse }

func (response GetDocument400JSONResponse) VisitGetDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetDocument401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetDocument401JSONResponse) VisitGetDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type GetDocument404JSONResponse struct{ NotFoundJSONResponse }

func (response GetDocument404JSONResponse) VisitGetDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type PrintDocumentRequestObject struct {
	DocumentUuid DocumentUUID `json:"document_uuid"`
}

type PrintDocumentResponseObject interface {
	VisitPrintDocumentResponse(w http.ResponseWriter) error
}

type PrintDocument204Response struct {
}

func (response PrintDocument204Response) VisitPrintDocumentResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PrintDocument400JSONResponse struct{ BadRequestJSONResponse }

func (response PrintDocument400JSONResponse) VisitPrintDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PrintDocument401JSONResponse struct{ UnauthorizedJSONResponse }

func (response PrintDocument401JSONResponse) VisitPrintDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type PrintDocument404JSONResponse struct{ NotFoundJSONResponse }

func (response PrintDocument404JSONResponse) VisitPrintDocumentResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create a new receipt
	// (POST /billing/documents/receipt)
	CreateReceipt(ctx context.Context, request CreateReceiptRequestObject) (CreateReceiptResponseObject, error)
	// Get a document by UUID
	// (GET /billing/documents/{document_uuid})
	GetDocument(ctx context.Context, request GetDocumentRequestObject) (GetDocumentResponseObject, error)
	// Print a document as HTML
	// (POST /billing/documents/{document_uuid}/print)
	PrintDocument(ctx context.Context, request PrintDocumentRequestObject) (PrintDocumentResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// CreateReceipt operation middleware
func (sh *strictHandler) CreateReceipt(ctx echo.Context) error {
	var request CreateReceiptRequestObject

	var body CreateReceiptJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateReceipt(ctx.Request().Context(), request.(CreateReceiptRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateReceipt")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateReceiptResponseObject); ok {
		return validResponse.VisitCreateReceiptResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetDocument operation middleware
func (sh *strictHandler) GetDocument(ctx echo.Context, documentUuid DocumentUUID) error {
	var request GetDocumentRequestObject

	request.DocumentUuid = documentUuid

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetDocument(ctx.Request().Context(), request.(GetDocumentRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetDocument")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetDocumentResponseObject); ok {
		return validResponse.VisitGetDocumentResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PrintDocument operation middleware
func (sh *strictHandler) PrintDocument(ctx echo.Context, documentUuid DocumentUUID) error {
	var request PrintDocumentRequestObject

	request.DocumentUuid = documentUuid

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PrintDocument(ctx.Request().Context(), request.(PrintDocumentRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PrintDocument")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PrintDocumentResponseObject); ok {
		return validResponse.VisitPrintDocumentResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}