# project/backend/billing/api/http/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Billing
  contact:
    name: Three Dots Labs
    url: https://threedots.tech/
servers:
  - url: https://{hostname}/billing
    variables:
      hostname:
        default: localhost
paths:
  /billing/documents/receipt:
    post:
      summary: Create a new receipt
      operationId: createReceipt
      tags:
        - documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocument'
      responses:
        '201':
          description: Receipt created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - document_uuid
                properties:
                  document_uuid:
                    $ref: '#/components/schemas/DocumentUUID'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /billing/documents/{document_uuid}:
    get:
      summary: Get a document by UUID
      operationId: getDocument
      tags:
        - documents
      parameters:
        - in: path
          name: document_uuid
          required: true
          schema:
            $ref: '#/components/schemas/DocumentUUID'
          description: UUID of the document to retrieve
      responses:
        '200':
          description: Document retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /billing/documents/{document_uuid}/print:
    post:
      summary: Print a document as HTML
      operationId: printDocument
      tags:
        - documents
      parameters:
        - in: path
          name: document_uuid
          required: true
          schema:
            $ref: '#/components/schemas/DocumentUUID'
          description: UUID of the document to print
      responses:
        '204':
          description: Document printed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    CreateDocument:
      type: object
      required:
        - issue_date
        - seller
        - buyer
        - currency
        - line_items
      properties:
        issue_date:
          type: string
          format: date-time
          description: Document issue date and time
        seller:
          $ref: '#/components/schemas/LegalEntity'
        buyer:
          $ref: '#/components/schemas/LegalEntity'
        currency:
          $ref: '#/components/schemas/Currency'
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
          minItems: 1
          description: Document line items

    LegalEntity:
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
          description: Legal entity name
        address:
          $ref: '#/components/schemas/Address'
        tax_id:
          $ref: '#/components/schemas/TaxID'

    Address:
      type: object
      required:
        - line1
        - line2
        - city
        - postal_code
        - country_code
      properties:
        line1:
          type: string
          description: First address line
        line2:
          type: string
          description: Second address line
        city:
          type: string
          description: City name
        postal_code:
          type: string
          description: Postal code
        country_code:
          type: string
          description: ISO 3166-1 alpha-2 country code
          x-go-type: shared.CountryCode
          x-go-type-import:
            path: eats/backend/common/shared

    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          $ref: '#/components/schemas/Decimal'
        currency:
          $ref: '#/components/schemas/Currency'

    Currency:
      type: string
      description: Currency code in ISO 4217 format
      x-go-type: shared.Currency
      x-go-type-import:
        path: eats/backend/common/shared

    LineItem:
      type: object
      required:
        - name
        - quantity
        - unit_amount
        - is_gross
      properties:
        name:
          type: string
          description: Item description/name
        quantity:
          type: integer
          minimum: 1
          description: Quantity of items
        unit_amount:
          $ref: '#/components/schemas/Decimal'
        is_gross:
          type: boolean

    DocumentUUID:
      type: string
      format: uuid
      description: UUID of a document
      x-go-type: domain.DocumentUUID
      x-go-type-import:
        path: eats/backend/billing/domain

    Decimal:
      type: string
      x-go-type: decimal.Decimal
      x-go-type-import:
        path: github.com/shopspring/decimal

    TaxID:
      type: string
      description: Tax identification number
      x-go-type: shared.TaxID
      x-go-type-import:
        path: eats/backend/common/shared

    DocumentType:
      type: string
      description: Type of the billing document
      enum:
        - receipt

    TaxType:
      type: string
      description: Type of tax applied
      enum:
        - vat
        - gst
        - sales-tax

    DocumentResponse:
      type: object
      required:
        - uuid
        - document_type
        - document_number
        - issue_date
        - currency
        - seller
        - buyer
        - line_items
        - summary
      properties:
        uuid:
          $ref: '#/components/schemas/DocumentUUID'
        document_type:
          $ref: '#/components/schemas/DocumentType'
        document_number:
          type: string
          description: Human-readable document number
        issue_date:
          type: string
          format: date-time
        currency:
          $ref: '#/components/schemas/Currency'
        seller:
          $ref: '#/components/schemas/LegalEntity'
        buyer:
          $ref: '#/components/schemas/LegalEntity'
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/ResponseLineItem'
        summary:
          $ref: '#/components/schemas/PriceBreakdownSummary'

    ResponseLineItem:
      type: object
      required:
        - name
        - quantity
        - breakdown
      properties:
        name:
          type: string
        quantity:
          type: integer
        breakdown:
          $ref: '#/components/schemas/PriceBreakdown'

    PriceBreakdown:
      type: object
      required:
        - tax_rate
        - tax_type
        - unit_net_amount
        - unit_tax_amount
        - unit_gross_amount
        - net_amount
        - tax_amount
        - gross_amount
      properties:
        tax_rate:
          $ref: '#/components/schemas/Decimal'
        tax_type:
          $ref: '#/components/schemas/TaxType'
        unit_net_amount:
          $ref: '#/components/schemas/Decimal'
        unit_tax_amount:
          $ref: '#/components/schemas/Decimal'
        unit_gross_amount:
          $ref: '#/components/schemas/Decimal'
        net_amount:
          $ref: '#/components/schemas/Decimal'
        tax_amount:
          $ref: '#/components/schemas/Decimal'
        gross_amount:
          $ref: '#/components/schemas/Decimal'

    PriceBreakdownSummary:
      type: object
      required:
        - net_amount
        - tax_amount
        - gross_amount
        - taxes
      properties:
        net_amount:
          $ref: '#/components/schemas/Decimal'
        tax_amount:
          $ref: '#/components/schemas/Decimal'
        gross_amount:
          $ref: '#/components/schemas/Decimal'
        taxes:
          type: array
          items:
            $ref: '#/components/schemas/TaxSummary'

    TaxSummary:
      type: object
      required:
        - tax_rate
        - tax_type
        - net_amount
        - tax_amount
      properties:
        tax_rate:
          $ref: '#/components/schemas/Decimal'
        tax_type:
          $ref: '#/components/schemas/TaxType'
        net_amount:
          $ref: '#/components/schemas/Decimal'
        tax_amount:
          $ref: '#/components/schemas/Decimal'

    ErrorResponse:
      type: object
      required:
        - slug
        - message
        - details
      properties:
        slug:
          type: string
          description: Error slug
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorDetails'

    ErrorDetails:
      type: object
      required:
        - error_slug
        - message
      properties:
        entity_type:
          type: string
        entity_id:
          type: string
        error_slug:
          type: string
        message:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'